handle _admin _feature only

+++ lib/feature/groups/group_setting/presentation/screen/group_setting_screen.dart
@@ -7,6 +7,7 @@ import 'package:telegram/core/component/clogo_loader.dart';
 import 'package:telegram/core/component/csnack_bar.dart';
 import 'package:telegram/core/component/general_image.dart';
 import 'package:telegram/core/di/service_locator.dart';
+import 'package:telegram/core/local/hive.dart';
 import 'package:telegram/core/routes/app_router.dart';
 import 'package:telegram/core/utililes/app_colors/app_colors.dart';
 import 'package:telegram/core/utililes/app_enum/app_enum.dart';
@@ -26,6 +27,7 @@ class GroupSettingsScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final user_id = HiveCash.read(boxName: 'register_info', key: 'id');
     return Scaffold(
       appBar: CAppBar(
         leadingIcon: Icons.arrow_back,
@@ -34,21 +36,51 @@ class GroupSettingsScreen extends StatelessWidget {
         },
         actions: [
           IconButton(
-            icon: const Icon(Icons.edit),
+            icon: const Icon(Icons.exit_to_app),
             onPressed: () {
-              // Navigate to Edit Group Screen
+              if (user_id ==
+                      sl<GroupCubit>()
+                          .state
+                          .members
+                          .firstWhere((element) => element.role == 'admin')
+                          .userId &&
+                  sl<GroupCubit>().state.members.length == 1) {
+                CSnackBar.showErrorDialog(context,
+                    'Group will be deleted if you did not set another admin',
+                    () {
+                  sl<GroupCubit>().deleteGroup(groupId);
+
+                  GoRouter.of(context).pop();
+                });
+                return;
+              } else {
+                CSnackBar.showErrorDialog(
+                    context, 'are you sure you want to leave group', () {
+                  sl<GroupCubit>().leaveGroup(groupId,
+                      HiveCash.read(boxName: 'register_info', key: 'id'));
+
+                  GoRouter.of(context).pop();
+                });
+              }
+              GoRouter.of(context).pop();
             },
           ),
-          IconButton(
-            icon: const Icon(Icons.delete),
-            onPressed: () {
-              CSnackBar.showErrorDialog(
-                  context, 'Are you sure you want to delete this group?', () {
-                sl<GroupCubit>().deleteGroup(groupId);
-                GoRouter.of(context).pop();
-              });
-            },
-          ),
+          if (sl<GroupCubit>()
+                  .state
+                  .members
+                  .firstWhere((element) => element.role == 'admin')
+                  .userId ==
+              user_id)
+            IconButton(
+              icon: const Icon(Icons.delete),
+              onPressed: () {
+                CSnackBar.showErrorDialog(
+                    context, 'Are you sure you want to delete this group?', () {
+                  sl<GroupCubit>().deleteGroup(groupId);
+                  GoRouter.of(context).pop();
+                });
+              },
+            ),
         ],
       ),
       body: BlocBuilder<GroupCubit, GroupState>(
@@ -111,36 +143,57 @@ class GroupSettingsScreen extends StatelessWidget {
                   sl<GroupCubit>().toggleNotifications(groupId, value);
                 },
               ),
-              const Divider(),
-              SwitchListTile(
-                activeColor: AppColors.primaryColor,
-                activeTrackColor: AppColors.primaryColor.withOpacity(0.5),
-                inactiveThumbColor: AppColors.grey,
-                inactiveTrackColor: AppColors.grey.withOpacity(0.5),
-                title: Text(
-                  'make it private',
-                  style: Theme.of(context).textTheme.bodyMedium,
+              if (user_id ==
+                  sl<GroupCubit>()
+                      .state
+                      .members
+                      .firstWhere((element) => element.role == 'admin')
+                      .userId)
+                Column(
+                  children: [
+                    const Divider(),
+                    SwitchListTile(
+                      activeColor: AppColors.primaryColor,
+                      activeTrackColor: AppColors.primaryColor.withOpacity(0.5),
+                      inactiveThumbColor: AppColors.grey,
+                      inactiveTrackColor: AppColors.grey.withOpacity(0.5),
+                      title: Text(
+                        'make it private',
+                        style: Theme.of(context).textTheme.bodyMedium,
+                      ),
+                      value: state.group!.privacy, //TODO: Get value from cubit
+                      onChanged: (value) {
+                        sl<GroupCubit>().togglePrivacy(groupId, value);
+                      },
+                    ),
+                  ],
                 ),
-                value: state.group!.privacy, //TODO: Get value from cubit
-                onChanged: (value) {
-                  sl<GroupCubit>().togglePrivacy(groupId, value);
-                },
-              ),
-              const Divider(),
-              Padding(
-                padding: const EdgeInsets.symmetric(vertical: 8.0),
-                child: TextButton.icon(
-                  onPressed: () {
-                    GoRouter.of(context)
-                        .push(AppRouter.kAddmembers, extra: state.group);
-                  },
-                  icon: const Icon(Icons.add, color: AppColors.primaryColor),
-                  label: const Text(
-                    'Add Member',
-                    style: TextStyle(color: AppColors.primaryColor),
-                  ),
+              if (user_id ==
+                  sl<GroupCubit>()
+                      .state
+                      .members
+                      .firstWhere((element) => element.role == 'admin')
+                      .userId)
+                Column(
+                  children: [
+                    const Divider(),
+                    Padding(
+                      padding: const EdgeInsets.symmetric(vertical: 8.0),
+                      child: TextButton.icon(
+                        onPressed: () {
+                          GoRouter.of(context)
+                              .push(AppRouter.kAddmembers, extra: state.group);
+                        },
+                        icon: const Icon(Icons.add,
+                            color: AppColors.primaryColor),
+                        label: const Text(
+                          'Add Member',
+                          style: TextStyle(color: AppColors.primaryColor),
+                        ),
+                      ),
+                    ),
+                  ],
                 ),
-              ),
               const Divider(),
               // Members List
               Column(
@@ -161,65 +214,74 @@ class GroupSettingsScreen extends StatelessWidget {
                               Theme.of(context).textTheme.bodySmall!.copyWith(
                                     color: AppColors.grey,
                                   )),
-                      trailing: PopupMenuButton<String>(
-                        color: const Color.fromARGB(255, 198, 217, 238),
-                        onSelected: (value) {
-                          if (value == 'editPermissions') {
-                            GoRouter.of(context)
-                                .push(AppRouter.kUserPermission, extra: member);
-                          } else if (value == 'remove') {
-                            sl<GroupCubit>()
-                                .removeMember(groupId, member.userId);
-                          } else if (value == 'admin') {
-                            sl<GroupCubit>().updateMemberRole(
-                              MemberModel(
-                                  userId: member.userId,
-                                  role: 'admin',
-                                  hasDownloadPermissions:
-                                      member.hasDownloadPermissions,
-                                  hasMessagePermissions:
-                                      member.hasMessagePermissions),
-                            );
-                          }
-                        },
-                        itemBuilder: (context) => [
-                          const PopupMenuItem(
-                              value: 'editPermissions',
-                              child: Row(
-                                children: [
-                                  Icon(Icons.edit),
-                                  SizedBox(
-                                    width: 5,
-                                  ),
-                                  Text('Edit Permissions'),
-                                ],
-                              )),
-                          const PopupMenuItem(
-                              value: 'remove',
-                              child: Row(
-                                children: [
-                                  Icon(Icons.delete),
-                                  SizedBox(
-                                    width: 5,
-                                  ),
-                                  Text('Remove Member'),
-                                ],
-                              )),
-                          const PopupMenuItem(
-                              value: 'admin',
-                              child: Row(
-                                children: [
-                                  Icon(Icons.star),
-                                  SizedBox(
-                                    width: 5,
-                                  ),
-                                  Text('set admin'),
-                                ],
-                              )),
-                        ],
-                      ),
+                      trailing: user_id ==
+                              sl<GroupCubit>()
+                                  .state
+                                  .members
+                                  .firstWhere(
+                                      (element) => element.role == 'admin')
+                                  .userId
+                          ? PopupMenuButton<String>(
+                              color: const Color.fromARGB(255, 198, 217, 238),
+                              onSelected: (value) {
+                                if (value == 'editPermissions') {
+                                  GoRouter.of(context).push(
+                                      AppRouter.kUserPermission,
+                                      extra: member);
+                                } else if (value == 'remove') {
+                                  sl<GroupCubit>()
+                                      .removeMember(groupId, member.userId);
+                                } else if (value == 'admin') {
+                                  sl<GroupCubit>().updateMemberRole(
+                                    MemberModel(
+                                        userId: member.userId,
+                                        role: 'admin',
+                                        hasDownloadPermissions:
+                                            member.hasDownloadPermissions,
+                                        hasMessagePermissions:
+                                            member.hasMessagePermissions),
+                                  );
+                                }
+                              },
+                              itemBuilder: (context) => [
+                                const PopupMenuItem(
+                                    value: 'editPermissions',
+                                    child: Row(
+                                      children: [
+                                        Icon(Icons.edit),
+                                        SizedBox(
+                                          width: 5,
+                                        ),
+                                        Text('Edit Permissions'),
+                                      ],
+                                    )),
+                                const PopupMenuItem(
+                                    value: 'remove',
+                                    child: Row(
+                                      children: [
+                                        Icon(Icons.delete),
+                                        SizedBox(
+                                          width: 5,
+                                        ),
+                                        Text('Remove Member'),
+                                      ],
+                                    )),
+                                const PopupMenuItem(
+                                    value: 'admin',
+                                    child: Row(
+                                      children: [
+                                        Icon(Icons.star),
+                                        SizedBox(
+                                          width: 5,
+                                        ),
+                                        Text('set admin'),
+                                      ],
+                                    )),
+                              ],
+                            )
+                          : Container(),
                     );
-                  }).toList(),
+                  }).toList()
                 ],
               ),
             ],
